\name{g}
\alias{g}
\title{g function}
\description{This function computes value of g function for given v,b,k,x,z}
\usage{g(v,b,k,x,z)}
\arguments{
	\item{v}{Number of treatments}
	\item{b}{Number of blocks}
	\item{k}{Block size}
	\item{x}{x}
	\item{z}{z}
	}
\value{
	\item{value}{Value of g function}
       }
\author{B N Mandal <mandal.stat@gmail.com>}
\keyword{internal}
