\name{gts}
\alias{gts}
\title{gts Value}
\description{Computes minimum value of g function for over permissible values of x,z for given v,b,k}
\usage{gts(v,b,k)}
\arguments{
	\item{v}{number of treatments}
	\item{b}{number of blocks}
	\item{k}{block size}	
	}
\value{
	\item{ming}{Minimum value of g function}
       }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}
