\name{interchange.NNP}
\alias{interchange.NNP}
\title{Interchange the Elements of a Concurrence Matrix to Get Highest Possible Lower Bound to A-efficiency}
\description{Interchanges the distinct off-diagonal elements of a given concurrence
 matrix and produces a concurrence matrix which has highest A-efficiency}
\usage{interchange.NNP(b,k,NNP)}
\arguments{
	\item{b}{number of blocks}
	\item{k}{block size}
	\item{NNP}{concurrence matrix}
	}
\value{
	A v x v matrix
       }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}
