\name{is.wholenumber}
\alias{is.wholenumber}
\title{Whole Number or Not}
\description{Checks whether a given number is whole number or not.}
\usage{is.wholenumber(x, tol = .Machine$double.eps^0.5)}
\arguments{
	\item{x}{a number}
	\item{tol}{tolerance. Default is tol = .Machine$double.eps^0.5}
	}
\value{Returns TRUE or FALSE}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}