\name{A_eff.NNP}
\alias{A_eff.NNP}
\title{A-efficiency of a proper binary incomplete block design}
\description{This function computes lower bound to A-efficiency of a binary proper incomplete 
block design given the block size, number of blocks and concurrence matrix.}
\usage{A_eff.NNP(b,k,NNP)}
\arguments{
	\item{b}{number of blocks}
	\item{k}{block size}
	\item{NNP}{concurrence matrix}
	}
\value{
	\item{Aeff}{A-efficiency}
       }
\author{B N Mandal <mandal.stat@gmail.com>}
\keyword{internal}
