\name{check.orthogonality}
\alias{check.orthogonality}
\title{Orthogonality of Rows of a Given Matrix}
\description{Checks whether rows of a given matrix are pairwise orthogonal or not. 
Returns a value 1 if the rows are pairwise orthogonal, else return 0.}
\usage{check.orthogonality(M)}
\arguments{
	\item{M}{a matrix}
	}
\value{
	A value of 1 if rows of the matrix are pairwise orthogonal else 0.
       }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}