\name{do.exist.NBIB}
\alias{do.exist.NBIB}
\title{Existence of a nearly balanced incomplete block design given the parameters}
\description{This function checks whether a nearly balanced incomplete block design exists or not 
for given parameters}
\usage{do.exist.NBIB(v,b,k)}
\arguments{
	\item{v}{number of treatments}
	\item{b}{number of blocks}
	\item{k}{block size}
	}
\value{Returns 1 if an NBIB design exists else return 0}
\author{B N Mandal <mandal.stat@gmail.com>}
\keyword{internal}