\name{g}
\alias{g}
\title{g Function}
\description{Computes value of g function for given v,b,k,x,z}
\usage{g(v,b,k,x,z)}
\arguments{
	\item{v}{number of treatments}
	\item{b}{number of blocks}
	\item{k}{block size}
	\item{x}{x}
	\item{z}{z}
	}
\value{
	\item{value}{Value of g function}
       }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}
