\name{gts}
\alias{gts}
\title{gts value}
\description{This function computes minimum value of g function for over permissible values of x,z for given v,b,k}
\usage{gts(v,b,k)}
\arguments{
	\item{v}{Number of treatments}
	\item{b}{Number of blocks}
	\item{k}{Block size}	
	}
\value{
	\item{ming}{Minimum value of g function}
       }
\author{B N Mandal <mandal.stat@gmail.com>}
\keyword{internal}
