\name{interchange.NNP}
\alias{interchange.NNP}
\title{Interchange the elements of a concurrence matrix to get highest possible A-efficiency}
\description{This function interchanges the distinct off-diagonal elements of a given concurrence
 matrix and produces a concurrence matrix which has highest A-efficiency}
\usage{interchange.NNP(b,k,NNP)}
\arguments{
	\item{b}{number of blocks}
	\item{k}{block size}
	\item{NNP}{concurrence matrix}
	}
\value{
	A v x v matrix
       }
\author{B N Mandal <mandal.stat@gmail.com>}
\keyword{internal}
