\name{is.wholenumber}
\alias{is.wholenumber}
\title{Whole number or not}
\description{This function checks whether a given number is whole number or not.}
\usage{is.wholenumber(x, tol = .Machine$double.eps^0.5)}
\arguments{
	\item{x}{A number}
	\item{tol}{Tolerance. Default is tol = .Machine$double.eps^0.5}
	}
\value{Returns TRUE or FALSE}
\author{B N Mandal <mandal.stat@gmail.com>}
\keyword{internal}